backend/requirements.txt
# FastAPI & Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Jinja2==3.1.2

# 数据库 & ORM
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9 # For PostgreSQL
alembic==1.12.1

# 认证 & 安全
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Pydantic for data validation
pydantic==2.5.2
pydantic-settings==2.1.0

# RAG & AI
chromadb==0.4.18
sentence-transformers==2.2.2
langchain==0.0.345
pypdf==3.17.1
docx2txt==0.8
unstructured==0.11.2
# Note: 'unstructured' may require additional system dependencies like 'libmagic'
# For markdown support in unstructured
markdown-it-py[linkify,plugins]==3.0.0

# 其他
requests==2.31.0
tqdm==4.66.1
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
scipy==1.11.4
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
transformers==4.35.2
tokenizers==0.15.0
tiktoken==0.5.1
httpx==0.25.1
websockets==12.0
openai==1.3.8
chardet==5.2.0
lxml==4.9.3
beautifulsoup4==4.12.2 